Ensure shared_ptr to the handler stays alive
If the parent thread destructor fires and no other copy of the
shared_ptr exists, then it'll try to clean up the HandlerQueue
object.  When that occurs, the condition variable will be cleaned
up; if a thread is waiting on the condition variable at that time,
then undefined behavior occurs (we've observed this translating
into a deadlock).