Work around an apparent timeout bug in libcurl on RHEL7
On RHEL7, we observed libcurl not setting an internal timeout during
DNS lookups; this caused the only timeout to be the periodic reporting
timeout every 5 seconds.  Depending on the outcome of the race, this
meant an operation might take 5 seconds (it appears to occur more often
in containers), resulting in extremely slow downloads.

Detect this issue by querying libcurl for the next timeout.  If there is
none -- but we know there is an ongoing operation (which should always
have a timeout!) -- switch to polling every 50ms.