First pass at using the connection broker
This allows a parent process to configure a Unix domain socket where
connected sockets can be passed.  The intent is the pelican process
will take the request, do a connection reversal, and pass a connected
socket to the origin back to the plugin.

The socket request is handled in the main handler loop and is non-blocking.