Provide an error code on metadata lookup failure
This provides an error code to prevent XRootD from inserting a nonsensical
one.

With this code - and a future patch to XRootD - the cache will produce a
"502 Bad Gateway" HTTP status on failure to determine an origin or an
offline origin.  This will allow much-improved error messages in the
Pelican client compared to today's generic "500 Internal Server Error".