Detect HTTP status code-based failures in curl worker loop
When the HTTP status code processing was done in the header callback,
it would call `Fail` there (triggering XrdCl callback) and then
`Success` was subsequently called by the curl worker (as the HTTP
request itself did succeed from curl's point-of-view).  That caused
the XrdCl callback to be invoked twice.  Depending on the outcome of
the race, this could cause a segfault if the handler object was
deleted after the initial failure.