Change CacheEntry structure to have pointer to self
The compiler on RHEL9 complained about having the `m_subdirs` member
of the `CacheEntry` class contain a copy of a `CacheEntry` itself
(complaining about only having a forward dec'l of `CacheEntry`).

The Ubuntu24 compiler was able to figure this out -- the `unordered_map`
doesn't need the size of the entry.  However, given these are copied
around when the data structure changes, switching to a reference to the
object is probably a more effective data structure regardless.