Permit tinyxml2 to come from a local tarball
On the OSG Koji builder, the version of tinyxml2 available for EL8
is too old.  The `ExternalProject_Add` command with a HTTPS URL is
not usable because Koji builds are not allowed access to the network
to aid reproducability.

This allows the builder to drop the tinyxml2 tarball in the
`cmake/tinyxml2` directory (particularly, it must be named
`tinyxml2-10.0.0.tar.gz`) to have that used instead of a download.