Don't look for `prefix` param in broker URL
We were previously looking for the `prefix` param
in the broker URL set in the `x-pelican-broker`
header. This param was then sent to the cache Go
process to connect to the broker through the broker.
However, this is no longer required. All communication
channels are now origin-host specific and no longer
tied to the prefix.