Implement VectorRead functionality
The VectorRead method can be invoked by XCache when it is in "bypass"
mode; without this implementation, the cache will return "function
not implemented" under heavy load.

The majority of the implementation has gone into the parser for the
mime multipart byte range encoding; modest unit tests have been
included.