Switch queuing system to be based on `shared_ptr`
For writes, `Pelican::File` needed to keep a reference to the
in-progress operation for invocation from a future `Write` method.

However, on failure, since the curl worker owned the `unique_ptr`,
the memory was deleted and the subsequent attempt to use it from
the `File` object caused a use-after-free.

Instead, the `shared_ptr` will just decrement the use count and
the `File` object can safely invoke operations from outside the
curl worker thread.