Avoid mutating thread-unsafe state after invoking callback
The member data in a `CurlOperation` can be safely mutated from
a curl worker thread _until_ the handler is invoked.  At that point,
the client thread may be woken up and start making changes as well,
leading to race conditions.

The old behavior of setting `m_handler = nullptr;` after the callback
was invoked had caused race conditions observed in the wild:

- T1: `HandleResponse` was invoked from the curl thread.
- T2: Client thread woke up and was given additional data via the
  `Write` method, triggering a continuation of the `CurlPutOp`
  operation.
- T2: `CurlPutOp` was added to the operation queue and `m_handle` set.
- T1: The curl thread sets `m_handle = nullptr`.
- T1: The curl thread continues the operation with a null callback,
  and hence doesn't invoke the callback when done.

When this race condition occurred, the write callback is never invoked
and it appears to the application that the write lasted an infinite
amount of time.