Handle 404 when creating a file
When opening a file that does not (yet) exist, we should initially
stat the object (as it might be a directory -- in that case we should
return an EISDIR).

If the stat fails with a 404 and we requested creating the file, then
we should intercept the failure and proceed regardless.