Replace explicit is_pelican code with OPTIONS check
Check for allowed verbs with the OPTIONS request; if PROPFIND is in
the list of allowed verbs, then use that for stat; otherwise, use
HEAD.

This is the proper thing to do for a generic HTTP server -- and it
happens to have the correct behavior around the Pelican director.
The director returns a 405 on the OPTIONS request, causing the new
code to use a HEAD against the director which is the correct thing
to do for Pelican.

This subsequently allows us to remove Pelican-specific code paths,
a step in preparing for an ultimate split between the generic HTTP
code and Pelican-specific code.