Clear the HTTP header setting before calling parent method
The `CurlOperation::ReleaseHandle` implementation resets the `m_curl`
handle.  If the parent method is called first here, then `m_curl`
is set to `nullptr` and we never cleanup the HTTP header.

Subsequently, on reuse of the handle, the pointer to the header list
is invalid and can cause an invalid memory read (potentially crashing
the process).