Correctly test for failed callback in the write function
If the write function fails the callback, the `CURLE_ABORTED_BY_CALLBACK`
error code is not given but rather `CURLE_WRITE_ERROR`.  Previously,
the `CURLE_WRITE_ERROR` fell through to the `else`, causing error
code 23 to pass back to the client.

Unfortunately, errno 23 is "Too many open files in the system" which
is highly misleading for operators.