Fix resource name confusion in the filesystem object
When passing from the "Pelican" to the "Curl" layer, we need to sometimes
use the resource name from the director (`/api/v1.0/...`) and not
the object name.  This corrects prior behavior when the object name was
used.

Without the fix, a director set to `stat` the object before responding
could get caught in a loop where a query from the cache to the director
is interpreted as a query for a cache and sent back to the same cache.