Fix lifetime issues for default callbacks
The put and prefetch operation, since they are not tied to a specific
File operation, may outlive the File object.  Thus, they need to
be able to live independently (here, we use a shared_ptr) and own
any memory they need to touch.

This caused loggers to migrate inside the object and the prefetch
state (enabled/disabled) to move inside the prefetch handler.