Fix parsing of `depth` header
Previously, if the director responded with invalid depth settings
(basically, larger than the actual depth of the URL), then the worker
thread would raise an exception.  This could kill off the running
queries in that thread, potentially blocking an open call (and causing
that file to stall until the cache service is restarted).