Implement Mkdir and Rmdir methods
This implements `mkdir` by creating an empty "sentinel" object that
causes the directory existence check logic to believe a directory
exists there.  `rmdir` simply removes a sentinel, if present.

This should be sufficient for some simple cases (like FTS3 transfers)
but a far cry from POSIX.  For example, removing the sentinel will be
successful when deleting an empty directory -- but a non-empty directory
will continue to exist even after a successful `rmdir` operation.