Do not initialize plugin until it is used
This defers the initialization of the plugin to first-use instead of
at load.  XrdCl tends to load all plugins available, meaning the thread
pool always was initialized in anything that used XRootD's client.

This improves the fork-safety of the plugin: the plugin will function
post-fork as long as it's never used in the parent.  This enables unit
tests to fork and then run the tests.