Fix memory leaks in the director cache open handler
This utilizes unique_ptr's (like we do elsewhere in the code base)
to ensure memory is freed across multiple tricky paths and, for the
case of an "open" request, makes sure the final handler is invoked
with the `nullptr` for the return object (which is what they usually
expect).  Some handlers do not delete a response if provided with
one because they don't expect them, causing a memory leak.