Ensure callers get the correct response object
The client caller invoking the pelican plugin expects the standard
XRootD response objects.  When we get the extended response objects
back internally, we must convert those to the standard object if
the standard object doesn't have a virtual destructor.

Otherwise, the incorrect destructor is invoked on free and potentially
frees the wrong amount of memory.  Fixes an ASAN-flagged issue.