Do not allow RPM specfile to be set to an existing tag
To prevent developers from forgetting to maintain the RPM specfile,
force all PRs and pushes to fail unit tests if the RPM's version was
not bumped to a new values after the last release.

This ensures all pre-release testing is done with the upcoming version
number, not the previous one.