Add client config for querying for cache, not origin
If a Pelican cache wants to run without registration, it cannot
contact origins directly (only registered caches can get an appropriate
credential to do so).  Instead, it must act as a client and ask the
director for other caches to leverage.

This commit adds such a mode and corresponding unit test.