Ensure lock is not held when invoking the read operation
The prefetch constructor will invoke the read operation's `Continue`
method.  If that fails, the callback will be invoked with the failure.

Unfortunately, the callback itself wants to grab the same mutex that's
held by the constructor, resulting in a deadlock.

This commit refactors the constructor so it optionally takes a copy
of the lock and will unlock/lock it around the invocation of the
read operation.  If the operation fails, the caller must catch the
exception and handle it appropriately.