Redirect stdout/err of the death test to parent
We've been having failures that appear to be missing relevant information
from stdout/err in the child.  Rewire the descriptors to a pipe available
to the parent.  The parent buffers the text from the child and sends it
back down to the child; the latter writes it to the descriptor expected
by gtest.